python/requirements.txt
fastapi
uvicorn[standard]
discord.py
aiosqlite
python-multipart